move Input to original location
