+++ src/Components/Input/react-mentions/utils/applyChangeToValue.js
