+++ src/Components/Input/react-mentions/utils/countPlaceholders.js
