+++ src/Components/Input/react-mentions/utils/diacritics.js
