+++ src/Components/Input/react-mentions/utils/escapeRegex.js
