+++ src/Components/Input/react-mentions/utils/findPositionOfCapturingGroup.js
