+++ src/Components/Input/react-mentions/utils/findStartOfMentionInPlainText.js
