+++ src/Components/Input/react-mentions/utils/getEndOfLastMention.js
