+++ src/Components/Input/react-mentions/utils/getPlainText.js
