+++ src/Components/Input/react-mentions/utils/isIE.js
