+++ src/Components/Input/react-mentions/utils/isNumber.js
