+++ src/Components/Input/react-mentions/utils/isPlainObject.js
