+++ src/Components/Input/react-mentions/utils/mapPlainTextIndex.js
