+++ src/Components/Input/react-mentions/utils/readConfigFromChildren.js
