+++ src/Components/Input/react-mentions/utils/spliceString.js
