+++ src/Components/Input/useTextTypingEffect.ts
