+++ src/utils/activeSocialInfoString.ts
@@ -1,4 +1,4 @@
-import { getAllWordsAndMentions } from '../Components/Search/Input/utils';
+import { getAllWordsAndMentions } from '../Components/Input/utils';
 
 export const getActiveSocialInfoString = ({
   dappName,
